+++ api/include/opentelemetry/metrics/noop.h
@@ -2,15 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
-#include "opentelemetry/metrics_new/async_instruments.h"
-#include "opentelemetry/metrics_new/meter.h"
-#include "opentelemetry/metrics_new/meter_provider.h"
-#include "opentelemetry/metrics_new/observer_result.h"
-#include "opentelemetry/metrics_new/sync_instruments.h"
-#include "opentelemetry/version.h"
+#ifndef ENABLE_METRICS_PREVIEW
+
+#  include "opentelemetry/metrics/async_instruments.h"
+#  include "opentelemetry/metrics/meter.h"
+#  include "opentelemetry/metrics/meter_provider.h"
+#  include "opentelemetry/metrics/observer_result.h"
+#  include "opentelemetry/metrics/sync_instruments.h"
+#  include "opentelemetry/version.h"
 
 OPENTELEMETRY_BEGIN_NAMESPACE
-namespace metrics_new
+namespace metrics
 {
 
 template <class T>
@@ -295,5 +297,6 @@ public:
 private:
   nostd::shared_ptr<Meter> meter_;
 };
-}  // namespace metrics_new
+}  // namespace metrics
 OPENTELEMETRY_END_NAMESPACE
+#endif
