+++ api/test/metrics/noop_sync_instrument_test.cc
@@ -1,14 +1,16 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-#include <gtest/gtest.h>
-#include <map>
-#include "opentelemetry/metrics_new/noop.h"
+#ifndef ENABLE_METRICS_PREVIEW
+
+#  include <gtest/gtest.h>
+#  include <map>
+#  include "opentelemetry/metrics/noop.h"
 
 TEST(Counter, Add)
 {
-  std::shared_ptr<opentelemetry::metrics_new::Counter<long>> counter{
-      new opentelemetry::metrics_new::NoopCounter<long>("test", "none", "unitless")};
+  std::shared_ptr<opentelemetry::metrics::Counter<long>> counter{
+      new opentelemetry::metrics::NoopCounter<long>("test", "none", "unitless")};
 
   std::map<std::string, std::string> labels = {{"k1", "v1"}};
   EXPECT_NO_THROW(counter->Add(10l, labels));
@@ -18,8 +20,8 @@ TEST(Counter, Add)
 
 TEST(histogram, Record)
 {
-  std::shared_ptr<opentelemetry::metrics_new::Histogram<long>> counter{
-      new opentelemetry::metrics_new::NoopHistogram<long>("test", "none", "unitless")};
+  std::shared_ptr<opentelemetry::metrics::Histogram<long>> counter{
+      new opentelemetry::metrics::NoopHistogram<long>("test", "none", "unitless")};
 
   std::map<std::string, std::string> labels = {{"k1", "v1"}};
   EXPECT_NO_THROW(counter->Record(10l, labels));
@@ -29,11 +31,13 @@ TEST(histogram, Record)
 
 TEST(UpDownCountr, Record)
 {
-  std::shared_ptr<opentelemetry::metrics_new::UpDownCounter<long>> counter{
-      new opentelemetry::metrics_new::NoopUpDownCounter<long>("test", "none", "unitless")};
+  std::shared_ptr<opentelemetry::metrics::UpDownCounter<long>> counter{
+      new opentelemetry::metrics::NoopUpDownCounter<long>("test", "none", "unitless")};
 
   std::map<std::string, std::string> labels = {{"k1", "v1"}};
   EXPECT_NO_THROW(counter->Add(10l, labels));
   EXPECT_NO_THROW(counter->Add(2l));
   EXPECT_NO_THROW(counter->Add(10l, {{"k1", "1"}, {"k2", 2}}));
 }
+
+#endif
